Migrate editor state hook specs to renderEntry
The `useContentElementCommandSubscription` and
`useContentElementEditorState` specs (`inFrontend` and `inEditorPreview`
variants) now go through `renderEntry` from the matching page objects
module instead of `renderInEntry(<Entry />)`. Single-element tests use
the new `contentElement: {ui: <Test />}` shorthand instead of
registering content element types separately; the specs move to
`frontend/features/` and `frontend/inlineEditing/features/`
respectively, so the mode is expressed by the file's location rather
than an `inFrontend` / `inEditorPreview` directory.
